requirements.txt
python-telegram-bot==20.3
gspread==5.7.2
oauth2client==4.1.3
